Moved methods into QRCollectionActivity
